New pages, platform optimisation, footer, content,
Added 4 new pages to the website, all having the same structure.  Added the same footer to every page. Wrote content for the new 4 pages, added images to all pages. Optimised the website for phones in portrait mode, tablets in portrait, tablets in landscape, and computer screens. It's not optimised for phones in landscape mode.